Attack speed rework, MaxLevel cap, dual-wield animation system
- AttackBaseMs: 1000 -> 2000, divisor: 5 -> 12 for smoother curve
- GetAttackSpeed/GetAttackSpeedWithWeapon return double (no int truncation)
- MaxLevel = 50 in Balance.cs, BalanceStatic.cs, Player.TryLevelUp()
- LH_Sword_SS.png added (24-row atlas for off-hand weapon sprites)
- animations.json: 80 entries, Rows 20->24 for Character/Sword/LH_Sword
- Dual-wield: main hand + second attack within single AttackIntervalMs
- Attack animation synced to server AttackIntervalMs
- TriggerAttack() no longer resets timer if already attacking
- SpriteCache: off-hand weapon animation getters added
- GameScreen: lhand weapon subtype wiring
- ClientMessages: AttackSpeed as double
- Known issue: doubled animation during dual-wield (to fix)

+++ RPGO.ClientMonoGame/Networking/ClientMessages.cs
@@ -57,7 +57,7 @@ public sealed class StatusData
     public int SkillPoints { get; set; }
     public int Speed { get; set; }
     public int MoveIntervalMs { get; set; }
-    public int AttackSpeed { get; set; }
+    public double AttackSpeed { get; set; }
     public int AttackIntervalMs { get; set; }
     public string WeaponDamageType { get; set; } = "";
     public double WeaponSpeedModifier { get; set; } = 1.0;
